all bugs fixed

+++ ITS-BRD/Programs/GS_Praktika_2/Src/zeitmessung.c
@@ -17,5 +17,5 @@ t1 = getTimeStamp();
 double getZeitspanne(){
 	t2 = t1;
 	t1 = getTimeStamp();
-	return t1 - t2;
+	return (uint32_t)t1 - t2;
 }
